Add LSP for C/C++ (clangd)
This commit also modifies the way we were specifying which LSP's to
auto-install. We were using vim.lsp.get_clients() before to query the
list of LSP's, but this would only return active clients, and since
these are loaded lazily, they wouldn't necessarily be active when neovim
starts up.

Instead, list the LSP's explicitly in configs.lspconfig and export it,
so it can be referenced in configs.mason-lspconfig.

Signed-off-by: Sanjay Narayanan <sanjaynarayanan94@gmail.com>